refactor(client): extract shared date/error/escape utilities
