+++ client/src/components/ConfirmModal.jsx
@@ -1,5 +1,5 @@
-import { useEffect } from 'react';
 import Button from './Button';
+import useEscapeKey from '../hooks/useEscapeKey';
 import styles from '../styles/ConfirmModal.module.css';
 
 // 위험·되돌릴 수 없는 액션 직전에 한 번 더 확인받는 모달
@@ -13,14 +13,7 @@ export default function ConfirmModal({
   onCancel,
 }) {
   // ESC로 취소
-  useEffect(() => {
-    if (!open) return;
-    const onKey = (e) => {
-      if (e.key === 'Escape') onCancel?.();
-    };
-    window.addEventListener('keydown', onKey);
-    return () => window.removeEventListener('keydown', onKey);
-  }, [open, onCancel]);
+  useEscapeKey(open, onCancel);
 
   if (!open) return null;
