+++ client/src/utils/date.js
@@ -1,0 +1,18 @@
+// Relative time, e.g. "today", "1d ago", "2w ago".
+export function formatRelative(iso) {
+  const diff = Date.now() - new Date(iso).getTime();
+  const days = Math.floor(diff / 86400000);
+  if (days === 0) return 'today';
+  if (days === 1) return '1d ago';
+  if (days < 7) return `${days}d ago`;
+  if (days < 14) return '1w ago';
+  return `${Math.floor(days / 7)}w ago`;
+}
+
+// Absolute timestamp, e.g. "2026-06-04 14:30".
+export function formatAbsolute(iso) {
+  if (!iso) return '';
+  const d = new Date(iso);
+  const pad = (n) => String(n).padStart(2, '0');
+  return `${d.getFullYear()}-${pad(d.getMonth() + 1)}-${pad(d.getDate())} ${pad(d.getHours())}:${pad(d.getMinutes())}`;
+}
